feat(cli): extract the lg CLI into an importable pkg/cli
The whole command line lived in package main at the module root, so a
third-party module could only build an lg-runtime-style bytecode host —
no repl, no resolver, no -b. That is not enough for a binary built around
generated interop, which has to BE lg: it runs scripts against the native
namespace, and it is the compiling side of lg -b, where top-level forms
execute at AOT time and a veneer's (:require) must resolve that namespace
at build time too.

Move lg.go and its build-tagged siblings into pkg/cli behind
Main(version, commit string) int, which returns an exit code instead of
calling os.Exit. The move is mechanical: the only edit to the tagged
files is the package clause.

Two things deliberately stay at the root. version/commit, because
goreleaser and the Makefile both set -X main.version / -X main.commit and
ldflags can only reach the declaring package — Main takes them as
arguments. And lg_gogen_ir.go, a gitignored build artifact of blank
imports whose path, package and tag are hardcoded in cmd/lgbgen/main.go,
the Makefile, scripts/fanout-ratchet.lg and a determinism test; blank
imports behave identically from package main, so moving it costs four
call sites and buys nothing.

Flag registration moves out of init() into a sync.Once from Main, so
importing pkg/cli no longer mutates the global flag set before a custom
main can register its own. Flags stay on flag.CommandLine rather than a
private FlagSet precisely so a custom main's own flags still get parsed;
the consequences — -h exits from inside Main, and Main is
call-once-per-process because option state is package-level — are stated
in its doc comment rather than implied away.

One bug this creates and fixes in the same breath: -w passed the
ldflags-stamped version to gomod.Generate, which turns it into
github.com/nooga/let-go@v<version>. Correct while let-go was always the
main module, wrong once pkg/cli is importable, since a custom binary's
stamp describes its OWN module — a custom lg at v2.0.0 would pin the
generated WASM module to a let-go release that does not exist. Read
let-go's entry from BuildInfo.Deps instead.

Also fixed while here: profile flags still registered from init() under
lg_profile, the exact import side effect the other flags were moved out
of init() to avoid.

The e2e builds the real thing — generate interop for a stdlib package,
blank-import it beside pkg/cli, run a script through the resulting
binary, then bundle with that binary as its own bundle base and run the
result. Hermetic: a replace directive and a stdlib target, so nothing is
fetched.

Writing it surfaced an emitter bug. ex-struct? tested the export vector's
LENGTH, but all three :type serialization branches produce six elements —
the non-struct branch writes `nil []` where the struct branch writes
`:struct [fields]` — so it answered true for every named type and emitted
vm.RegisterStruct for named arrays. hash/crc32's Table is
`type Table [256]uint32`, so the generated package panicked at init with
"expected struct, got array" before main ever ran. The in-tree xxh3
golden never caught it because xxh3 uses -opaque-structs, the one flag
that skips RegisterStruct. Test the marker instead; both directions
pinned; golden byte-identical.

docs/guide/custom-lg.md documents the module layout, the blank-import
pattern, the init-timing note and the bundle-base story, including that
-w cannot carry custom namespaces: the WASM build scaffolds a fresh
module from a fixed template that imports runtime packages only.

Verified as a pure refactor: make test green; default, lg_profile,
gogen_ir, plan9 and js/wasm all build; and bin/lg -v, -e and the full -h
flag set are byte-identical to a binary built from main, bar the usage
line's program path. GOOS=windows fails in pkg/rt/term.go before and
after — pre-existing.

Baseline: main a665761. Depends on the lginterop out-of-tree change.
