+++ docs/README.md
@@ -59,6 +59,7 @@ A subdir is earned when a cluster of related docs justifies one; one-off cross-c
 | Embedding let-go in a Go program | `guide/embedding-in-go.md` |
 | Dropping `net/http` from a build (`lg_no_http`) | `guide/embedding-in-go.md` (§ Dropping `net/http`) |
 | Wrapping Go packages (lginterop, `database/sql` example) | `guide/go-interop.md` |
+| Building a custom `lg` around your own Go packages (`pkg/cli`) | `guide/custom-lg.md` |
 | Running as a WASI (`GOOS=wasip1`) module | `guide/wasi.md` |
 | let-go under TinyGo (status: not currently supported) | `guide/tinygo.md` |
 | nREPL server + editor setup | `guide/nrepl.md` |
